Move api-stack to app plane.
