+++ cdk/src/open-data-platform/app-plane/app-plane-stack.ts
@@ -6,6 +6,7 @@ import { CommonProps } from '../../util';
 import { DataPlaneStack } from '../data-plane/data-plane-stack';
 import { NetworkStack } from '../network/network-stack';
 import { TileServer } from './tileserver/tileserver';
+import { ApiStack } from './api/api-stack';
 
 // TODO: consider narrowing the props down to the specific things that the App Plane needs.
 // E.g. just the cluster object itself, rather than the entire network stack.
@@ -16,6 +17,7 @@ interface AppPlaneStackProps extends CommonProps {
 
 export class AppPlaneStack extends Stack {
   readonly tileServer: TileServer;
+  readonly api: ApiStack;
 
   constructor(scope: Construct, id: string, props: AppPlaneStackProps) {
     super(scope, id, props);
@@ -26,6 +28,13 @@ export class AppPlaneStack extends Stack {
       ecsCluster: networkStack.cluster,
       postgisTileDatabase: dataPlaneStack.cluster,
       databaseUrlSecret: dataPlaneStack.tileserverCredentials.databaseUrlSecret,
+    });
+
+    this.api = new ApiStack(this, 'ApiStack', {
+      cluster: dataPlaneStack.cluster,
+      vpc: networkStack.vpc,
+      db: dataPlaneStack.databaseName,
+      credentialsSecret: dataPlaneStack.cluster.secret!,
     });
   }
 }
